requirements.txt
# Core
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0

# Database
SQLAlchemy>=1.4.0
psycopg2-binary>=2.9.0

# Airflow
apache-airflow==2.8.0
apache-airflow-providers-postgres>=5.0.0

# Dashboard
streamlit>=1.28.0
plotly>=5.18.0

# Environment
python-dotenv>=1.0.0